Added a PIL alpha_over extension in C, and code to setup.py to build it
I needed a way to let py2exe options be provided only when py2exe is
available. My solution (maybe not the best) was to put all the
arguments for the final setup() call into a dictionary, that is filled
in conditionally during the script's execution. This is why the file
is completely changed; however, it still does what it used to do (when
py2exe is available).