Add --changelist and --changelist-format
Allow output of changed tiles to a file specified with --changelist. The
file contains the name of each tile on a line. The path can be relative
or absolute, chosen using --changelist-format.

The code assumes that all innertiles will have been updated when a
worldtile changes, and prints the corresponding filenames to the list
directly, rather than add message passing to the innertile methods.