Reworked handling of bare consoles
Now the "Please run overviewer from the command line" message will
only be displayed in select cases.  This supports cases when users
want to give all the correct arguments to overviewer via a shortcut
or something similar.  In these cases, a 'Press [enter] to close
this window' message will still be displayed.