Fix tile range error in chunk scan
The chunk scan should have been checking tile bounds using the treedepth
value, not the world boundary. This was causing some tiles near the edge
to not get rendered. In the old version, the row/col boundaries were set
to the treedepth calculated bounds, so this wasn't a problem there.