Fixed some render-tiles not being rendered
If a render-tile is only touched by chunks with blocks at the highest
point near the sky, then the render tile may not be rendered due to the
chunk scan's algorithm. Chunks in rows divisible by 4 were not
considering the tile above them, even though the chunk partially
overlaps. This is only seen for chunks with blocks at the very top of
them by the sky, and only when no other chunks would trigger that tile
for rendering.