Fixed bug in chunkrenderer, other small changes
Compiled block definitions didn't keep a proper python reference to
Image objects, which puts them at risk for garbage collection. (Usually
the Textures object keeps a reference, but not if that is itself garbage
collected)
Now, the BlockDefs struct keeps a reference to a python list object,
which holds all images used as textures. This list is never used for
anything but making sure those images don't get garbage collected.

Other changes in this commit:
* changed bitflags in the BlockDef struct to unsigned
* cast py_render_block function pointer to PyCFunction to silence a
  warning