Explain importing for filter functions
This code is correct, but doesn't work correctly with signs since they got JSON text formats in 14w25a. I will try to create another PR with an improved sign function tomorrow, but it's turning out to be complex, so for the moment, I'll stick with this tip on making imports work.