Build and test rpm and deb repos during CI
Note that at the moment, this is building and testing a "test" repo.
This test repo is not the same one that you use by following our docs.
Once this new code is better tested, we'll switch over (existing users
will not have to do anything to use this new repo).