readme: add API key note, small fixes
We previously stated that a Google Maps API key is not required.
However, this is not true (anymore). I'm still wary of adding
it to the documentation since my hopes is that in the time it would
take me to add this to the docs I could just finish leaflet, but this
is the least we can do to clear up any confusion.

Also switched some URLs to use HTTPS, which closes #1314.

Lastly, fixed a warning about a not long enough title line.