world: work around minecraft palette bugs
Instead of trusting that Minecraft will only request up to index
num_palette_entries - 1, we'll assume Mojang occasionally produces
a weird file where it tries to request past that, up to the maximum
palette index it can based on the bits per value. To facilitate this,
this commit extends the translation palette array to this maximum size.

Such invalid palette picks will simply be replaced by ID 0 (aka air),
so it should cause minimal weirdness in the output.

Works around #1797.