setup: find system python path for CI hack
The problem appears to be that virtualenv overrides the base_prefix
so distutils can no longer find the system python path when adding
linker args.

This adds an ugly hack to find the system python path, and adds it
to the linker flags on Windows. Yay, no more hardcoded paths.

This should also make win32 builds work, and avoid adding linker
args manually.